reads: safe mismatch counters, parallel dedupe; 0.11.35 -> 0.11.36
Two pieces of work on the read path.

reads/ is now free of unsafe. Both unsafe SSE2 mismatch counters are replaced:
count_mismatches routes through bio's simd hamming, count_mismatches_bounded
becomes a 16-byte block loop with a scalar tail. The bounded contract is
preserved -- exact whenever the result is at or under the limit -- so no caller
observes a difference. Record gains a len field in the padding off and count
already forced, which makes id_to_slot dead; dropping it frees ~448 MB at the
112M-uniques scale and turns finalize's random bucket lookup into a sequential
record read. The dupes blob streams keys and values as they are found instead
of pre-reserving. Three latent overflow hazards are fixed (u8 quality
arithmetic wrapping above 222, compare_req underflow in polyx) plus a fourth
the verification fuzz found: polyx's walk-back underflows on a pure homopolymer
read, which panics in release too. 925,080 fuzz comparisons against the old
implementation, zero divergence.

dedupe's build_table now runs a pipeline: one reader thread per mate file, N
workers running the trim chain plus canonicalisation and hashing, and a
consumer that inserts. Inflate measured 84% of the reader's cost and the two gz
streams are independent, so each mate gets its own thread; the zipper pairs
batch k of one with batch k of the other, which keeps pairing and ordering
identical to reading them in lockstep.

The table insert stays single-threaded deliberately. next_id is assigned in
encounter order and the whole output depends on it, so batches are dealt to
workers round robin and taken back in the same round robin -- the consumer sees
survivors in exactly the serial order and ids match with no sorting or
renumbering. Workers only do work that is pure given TrimOptions; their scratch
and stats are private and merged at the end, which is sound because every
TrimStats field is a counter or a histogram.

SAPPHYRE_THREADS sets the worker count; 0 selects the original serial path,
kept as the reference the parallel one is validated against. The default is
min(cores - 2, 6): throughput plateaus once the workers can keep the insert
fed, and past six only peak RSS moves.

  DDM8637    20.18s -> 5.23s   3.85x   798 MB -> 880 MB
  DDM9435   204.37s -> 48.22s  4.24x  5.42 GB -> 5.48 GB

Memory overhead is bounded channel buffering, so it is fixed rather than
proportional -- +10.3% on the small library, +1.1% on the large one.

Byte-exact throughout: batches and packed_dupes sha256 identical to the
pre-change build on both libraries, at 0, 1, 2, 3, 5, 7, 8 and 14 workers;
across ragged mates, input below DETECT_SAMPLE, single-end and 3-pair input;
and through the table growth path. cargo test 100 passed.

+++ src/reads/filter.rs
@@ -139,7 +139,7 @@ pub fn pass_filter(r: &ReadRec<'_>, opts: &FilterOptions) -> FilterVerdict {
 
     if let Some(q) = opts.qual {
         if q.enabled && r.has_qual() {
-            let m = count_quality_metrics(r.quals(), r.bases(), q.qualified_qual + 33);
+            let m = count_quality_metrics(r.quals(), r.bases(), i64::from(q.qualified_qual) + 33);
             if (m.low_qual as f64) > q.unqualified_percent_limit * rlen as f64 / 100.0 {
                 return FilterVerdict::FailQuality;
             }
@@ -266,7 +266,7 @@ pub fn trim_and_cut(
             s += 1;
         }
         if found_low_qual_window {
-            while s < l - 1 && qual[s] >= 33 + cut.quality_right {
+            while s < l - 1 && i64::from(qual[s]) >= 33 + i64::from(cut.quality_right) {
                 s += 1;
             }
             rlen = s - front;
